[REFACTOR]: 패키지 위치 이동, AbstractAggregateRoot 추가
