+++ src/main/java/project/domain/workspace/domain/Workspace.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import project.common.exception.ErrorCode;
 import project.common.exception.PlantException;
-import project.domain.BaseEntity;
+import project.domain.common.BaseEntity;
 import project.domain.image.domain.Image;
 import project.domain.schedule.domain.Schedule;
 import project.domain.user.domain.User;
